Validations

+++ src/main/java/licos/bson/element/village/character/BsonAbstractCharacter.java
@@ -1,15 +1,17 @@
 package licos.bson.element.village.character;
 
+import dev.morphia.annotations.Id;
+import dev.morphia.annotations.Reference;
 import licos.bson.element.village.BsonName;
 import licos.bson.element.village.BsonElement;
 import lombok.Getter;
 import lombok.Setter;
 import org.bson.types.ObjectId;
-import org.mongodb.morphia.annotations.Id;
-import org.mongodb.morphia.annotations.Reference;
 
 public abstract class BsonAbstractCharacter extends BsonElement {
-    @Id @SuppressWarnings("unused")
+
+    @Id
+    @SuppressWarnings("unused")
     protected ObjectId _id;
 
     @Getter @Setter
@@ -19,7 +21,7 @@ public abstract class BsonAbstractCharacter extends BsonElement {
     protected String $id;
 
     @Getter @Setter
-    protected long id;
+    protected int id;
 
     @Getter @Setter @Reference
     protected BsonName name;
