+++ src/main/java/licos/bson/element/village/character/BsonStatusCharacter.java
@@ -1,16 +1,17 @@
 package licos.bson.element.village.character;
 
+import dev.morphia.annotations.Entity;
+import dev.morphia.annotations.Reference;
 import licos.bson.element.village.BsonName;
 import licos.bson.element.village.role.BsonSimpleRole;
 import licos.json.element.village.character.JsonStatusCharacter;
 import lombok.Getter;
 import lombok.Setter;
 import org.bson.types.ObjectId;
-import org.mongodb.morphia.annotations.Entity;
-import org.mongodb.morphia.annotations.Reference;
 
 @Entity("statusCharacters")
 public class BsonStatusCharacter extends BsonAbstractCharacter {
+
     @Getter @Setter @Reference
     private BsonSimpleRole role;
 
@@ -29,13 +30,12 @@ public class BsonStatusCharacter extends BsonAbstractCharacter {
     public BsonStatusCharacter(ObjectId _id,
                                String $context,
                                String $id,
-                               long id,
+                               int id,
                                BsonName name,
                                String image,
                                BsonSimpleRole role,
                                String status,
                                boolean isHumanPlayer) {
-                               //BsonAvatar avatar) {
         this._id = _id;
         this.$context = $context;
         this.$id = $id;
